moved filter function in useFilter-hook
